Schema Compare: keep Fabric Warehouse constraint scripts under table diffs
CreateDiffEntry stripped any child script that started with `alter` to avoid
duplicating constraint DDL that the parent's CREATE TABLE already inlined for
legacy SQL Server. SqlDwUnified (Fabric Warehouse) never inlines PK/FK/UNIQUE/
CHECK/DEFAULT constraints into CREATE TABLE; they are always emitted as
standalone `ALTER TABLE ... ADD CONSTRAINT ...` statements. The filter
therefore discarded the only script that defined the constraint, so the diff
editor showed CREATE TABLE without the PK and the generated script was missing
the PK definition entirely.

Detect constraint children (PK/FK/UNIQUE/CHECK/DEFAULT) under a SqlDwUnified
comparison and preserve their scripts in that case. Behavior is unchanged for
all other platforms.

Move the cached reflection-based platform detector from SchemaCompareOperation
into SchemaCompareUtils so CreateDiffEntry can use it; add a
ConditionalWeakTable cache keyed by SchemaComparisonResult so the recursive
walk pays the reflection cost once per comparison, not once per diff entry.
